requirements.txt
torch
torchvision
flask
pillow
numpy
tqdm
opencv-python
scikit-learn
